🚚 rename :: 컴포넌트 역할에 따라 template -> layout 으로 변경
