+++ src/refactoring/components/index.ts
@@ -1,4 +1,4 @@
-export * from './templates';
+export * from './layout';
 export * from './user';
 export * from './atoms';
 export * from './admin';
