+++ src/refactoring/components/layout/AppTemplate.tsx
