+++ src/refactoring/components/layout/SectionTemplate.tsx
